+++ CheckMyStar.Dal.Models/StarCriteriaDetail.cs
@@ -6,7 +6,7 @@ namespace CheckMyStar.Dal.Models
         public string StarLabel { get; set; } = string.Empty;
         public int CriterionId { get; set; }
         public string Description { get; set; } = string.Empty;
-        public decimal BasePoints { get; set; }
+        public int BasePoints { get; set; }
         public string TypeCode { get; set; } = string.Empty;
         public string TypeLabel { get; set; } = string.Empty;
     }
